+++ example/app/app/demos/custom-animations/advanced-parallax/index.tsx
@@ -1,6 +1,13 @@
 import * as React from "react";
 import { View } from "react-native";
-import Animated, { interpolate, interpolateColor, useAnimatedStyle } from "react-native-reanimated";
+import Animated, {
+  Extrapolation,
+  interpolate,
+  interpolateColor,
+  useAnimatedReaction,
+  useAnimatedStyle,
+  useSharedValue,
+} from "react-native-reanimated";
 import type { SharedValue } from "react-native-reanimated";
 import Carousel, { TAnimationStyle } from "react-native-reanimated-carousel";
 
@@ -8,6 +15,7 @@ import { SBItem } from "@/components/SBItem";
 import SButton from "@/components/SButton";
 import { ElementsText, window } from "@/constants/sizes";
 import { CaptureWrapper } from "@/store/CaptureProvider";
+import { scheduleOnRN } from "react-native-worklets";
 
 const PAGE_WIDTH = window.width;
 
@@ -15,6 +23,7 @@ interface ItemProps {
   index: number;
   animationValue: SharedValue<number>;
 }
+
 const CustomItem: React.FC<ItemProps> = ({ index, animationValue }) => {
   const maskStyle = useAnimatedStyle(() => {
     const backgroundColor = interpolateColor(
@@ -47,12 +56,19 @@ const CustomItem: React.FC<ItemProps> = ({ index, animationValue }) => {
     </View>
   );
 };
+
 function Index() {
   const [isAutoPlay, setIsAutoPlay] = React.useState(false);
-  const animationStyle: TAnimationStyle = React.useCallback((value: number) => {
+
+  const animationStyle: TAnimationStyle = React.useCallback((value: number, index: number) => {
     "worklet";
 
-    const zIndex = interpolate(value, [-1, 0, 1], [10, 20, 30]);
+    const zIndex = interpolate(
+      value,
+      value > 0 ? [0, 1] : [-1, 0],
+      value > 0 ? [10, 20] : [-10, 0],
+      Extrapolation.CLAMP
+    );
     const translateX = interpolate(value, [-2, 0, 1], [-PAGE_WIDTH, 0, PAGE_WIDTH]);
 
     return {
@@ -65,10 +81,9 @@ function Index() {
     <View style={{ flex: 1 }}>
       <CaptureWrapper>
         <Carousel
-          loop={true}
+          loop={false}
           autoPlay={isAutoPlay}
           style={{ width: PAGE_WIDTH, height: 240 }}
-          width={PAGE_WIDTH}
           data={[...new Array(6).keys()]}
           renderItem={({ index, animationValue }) => {
             return <CustomItem key={index} index={index} animationValue={animationValue} />;
@@ -77,6 +92,7 @@ function Index() {
           scrollAnimationDuration={1200}
         />
       </CaptureWrapper>
+
       <SButton
         onPress={() => {
           setIsAutoPlay(!isAutoPlay);