+++ example/app/app/demos/custom-animations/rotate-fade-in-out/demo.tsx
@@ -31,6 +31,7 @@ function Index() {
     <View
       id="carousel-component"
       dataSet={{ kind: "custom-animations", name: "rotate-fade-in-out" }}
+      style={{ width: window.width, height: 240, justifyContent: "center", alignItems: "center" }}
     >
       <Carousel
         loop
@@ -40,11 +41,11 @@ function Index() {
           justifyContent: "center",
           alignItems: "center",
         }}
-        width={PAGE_WIDTH}
-        height={PAGE_HEIGHT}
         data={[...new Array(6).keys()]}
         renderItem={({ index }) => {
-          return <SBItem key={index} index={index} />;
+          return (
+            <SBItem key={index} index={index} style={{ width: PAGE_WIDTH, height: PAGE_HEIGHT }} />
+          );
         }}
         autoPlay
         customAnimation={animationStyle}