+++ example/app/app/demos/custom-animations/rotate-in-out/demo.tsx
@@ -31,7 +31,11 @@ function Index() {
   }, []);
 
   return (
-    <View id="carousel-component" dataSet={{ kind: "custom-animations", name: "rotate-in-out" }}>
+    <View
+      id="carousel-component"
+      dataSet={{ kind: "custom-animations", name: "rotate-in-out" }}
+      style={{ width: window.width, height: 240, justifyContent: "center", alignItems: "center" }}
+    >
       <Carousel
         loop
         style={{
@@ -40,10 +44,12 @@ function Index() {
           justifyContent: "center",
           alignItems: "center",
         }}
-        width={PAGE_WIDTH}
-        height={PAGE_HEIGHT}
         data={[...new Array(6).keys()]}
-        renderItem={({ index }) => <SBItem key={index} index={index} />}
+        renderItem={({ index }) => {
+          return (
+            <SBItem key={index} index={index} style={{ width: PAGE_WIDTH, height: PAGE_HEIGHT }} />
+          );
+        }}
         autoPlay={AutoPLay.status}
         customAnimation={animationStyle}
       />